(refactor) move repo metadata GQL API resolvers to repository_metadata.go
